test: melhoria na organização dos testes
